[Feature] [+] = reworking the way entities are spawned taking now a time in float and changing the output from the number in the enum to the string of the config file

+++ src/ecs/Systems/HandleEntitySpawnSystem.hh
@@ -62,6 +62,6 @@ namespace RType {
 
             std::function<void(const std::basic_string<unsigned char> &message)> _sendToAllClient;
             std::unordered_map<EntityType, std::string> _entityTypeMap;
-            int _prevTime;
+            float _prevTime;
     };
 }
